packages/word-processing/requirements.txt
# Core NLP dependencies
spacy>=3.4.0
wordfreq>=3.0.0

# High-accuracy lemmatization (Stanford NLP)
stanza>=1.5.0

# Collocation extraction (statistical methods)
nltk>=3.8.0

# Russian morphology (specialized)
pymorphy3>=2.0.0

# Configuration management
PyYAML>=6.0
pydantic>=2.6.0

# Modern CLI with Rich UI (Typer includes Rich)
typer[all]>=0.12.0

# Interactive prompts with autocomplete
prompt-toolkit>=3.0.43

# Development and testing
pytest>=7.0.0
pytest-xdist>=3.0.0
black>=22.0.0
mypy>=1.0.0
